Visualization integration part 2
