+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/common/DataDimensionItemType.java
@@ -39,5 +39,6 @@ public enum DataDimensionItemType
     PROGRAM_INDICATOR,
     PROGRAM_DATA_ELEMENT,
     PROGRAM_ATTRIBUTE,
+    EXPRESSION,
     VALIDATION_RULE
 }
