+++ dhis-2/dhis-services/dhis-service-core/src/main/resources/org/hisp/dhis/common/DataDimensionItem.hbm.xml
@@ -56,7 +56,11 @@
     
       <many-to-one name="attribute" class="org.hisp.dhis.trackedentity.TrackedEntityAttribute"
         column="programattribute_attributeid" foreign-key="fk_datadimensionitem_programattribute_attributeid" />
+
     </component>
-    
+
+    <many-to-one name="expressionDimensionItem" class="org.hisp.dhis.expressiondimensionitem.ExpressionDimensionItem"
+                 column="expressiondimensionitemid" foreign-key="fk_datadimensionitem_expressiondimensionitemid" />
+
   </class>
 </hibernate-mapping>
