+++ dhis-2/dhis-services/dhis-service-core/src/main/resources/org/hisp/dhis/expression_dimensionitem.hibernate/ExpressionDimensionItem.hbm.xml
@@ -10,7 +10,7 @@
 
         <cache usage="read-write"/>
 
-        <id name="id" column="expressionid">
+        <id name="id" column="expressiondimensionitemid">
             <generator class="native"/>
         </id>
