+++ dhis-2/dhis-services/dhis-service-core/src/main/resources/org/hisp/dhis/visualization/hibernate/Visualization.hbm.xml
@@ -327,8 +327,5 @@
 
     <!-- Dynamic attribute values -->
     <property name="attributeValues" type="jsbAttributeValues"/>
-
-    <many-to-one name="expressionDimensionItem" class="org.hisp.dhis.expressiondimensionitem.ExpressionDimensionItem"
-                 column="expressiondimensionitemid" unique="true" not-null="false" cascade="all"/>
   </class>
 </hibernate-mapping>
