+++ dhis-2/dhis-support/dhis-support-db-migration/src/main/resources/org/hisp/dhis/db/migration/2.40/V2_40_2__Add_column_datadimensionitem.sql
@@ -1,0 +1,6 @@
+ALTER TABLE IF EXISTS datadimensionitem
+    ADD COLUMN expressiondimensionitemid bigint;
+ALTER TABLE IF EXISTS datadimensionitem ADD CONSTRAINT fk_datadimensionitem_expressiondimensionitemid FOREIGN KEY (expressiondimensionitemid)
+        REFERENCES expressiondimensionitem (expressiondimensionitemid) MATCH SIMPLE
+        ON UPDATE NO ACTION
+        ON DELETE NO ACTION;
