--- dhis-2/dhis-support/dhis-support-db-migration/src/main/resources/org/hisp/dhis/db/migration/2.40/V2_40_2__Add_column_to_visualization_table.sql
@@ -1,9 +1,0 @@
-ALTER TABLE visualization ADD COLUMN IF NOT EXISTS expressiondimensionitemid bigint;
-ALTER TABLE visualization ADD CONSTRAINT fk_visualization_expressiondimensionitem FOREIGN KEY (expressiondimensionitemid)
-REFERENCES public.expressiondimensionitem (expressionid) MATCH SIMPLE
-ON UPDATE NO ACTION
-ON DELETE NO ACTION;
-
-CREATE UNIQUE INDEX uq_visualization_expressiondimensionitemid
-    ON public.visualization USING btree
-    (expressiondimensionitemid ASC NULLS LAST);
